--- parsers/extract_page_3&4.py
@@ -1,1 +1,0 @@
-"""Extracts the text from page 3 & 4, which is the cover story."""
