--- send_requests.py
@@ -1,1 +1,0 @@
-"""Send requests to Wordpress API."""
